+++ src/lightning_habana/pytorch/strategies/single.py
@@ -129,6 +129,7 @@ class SingleHPUStrategy(SingleDeviceStrategy):
             from quantization_toolkit import habana_quantization_toolkit  # noqa
 
             habana_quantization_toolkit.finish_measurements(self.model)
+            htcore.quantization.hpu_teardown_inference_env()
         return super().on_test_end()
 
     def on_predict_end(self) -> None:
@@ -136,6 +137,7 @@ class SingleHPUStrategy(SingleDeviceStrategy):
             from quantization_toolkit import habana_quantization_toolkit  # noqa
 
             habana_quantization_toolkit.finish_measurements(self.model)
+            htcore.quantization.hpu_teardown_inference_env()
         return super().on_predict_end()
 
     @classmethod