+++ tests/test_pytorch/test_deepspeed.py
@@ -826,8 +826,18 @@ def test_hpu_deepspeed_fp8_training_accuracy(tmpdir, get_device_count, stage):
 
     precision_plugin_params_list = [
         ({"precision": "bf16-mixed"}),
-        ({"precision": "16-mixed"}),
-        ({"precision": "fp8", "replace_layers": True, "recipe": recipe.DelayedScaling()}),
+        pytest.param(
+            {"precision": "16-mixed"},
+            marks=pytest.mark.skipif(
+                HPUAccelerator.get_device_name() == "GAUDI", reason="fp16 supported on Gaudi2 and above"
+            ),
+        ),
+        pytest.param(
+            {"precision": "fp8", "replace_layers": True, "recipe": recipe.DelayedScaling()},
+            marks=pytest.mark.skipif(
+                HPUAccelerator.get_device_name() == "GAUDI", reason="fp8 supported on Gaudi2 and above"
+            ),
+        ),
     ]
 
     loss_list = []