+++ src/lightning_habana/pytorch/strategies/fsdp.py
@@ -14,12 +14,11 @@
 import logging
 from contextlib import contextmanager
 from datetime import timedelta
-from typing import TYPE_CHECKING, Any, Callable, Dict, Generator, List, Literal, Optional, Set, Type, Union
+from typing import TYPE_CHECKING, Any, Callable, Generator, List, Literal, Optional, Set, Type, Union
 
 import torch
 from lightning_utilities import module_available
 from torch.nn import Module
-from torch.optim import Optimizer
 from typing_extensions import override
 
 if module_available("lightning"):
@@ -35,7 +34,7 @@ if module_available("lightning"):
     from lightning.fabric.utilities.types import ReduceOp
     from lightning.pytorch.plugins.precision import Precision
     from lightning.pytorch.strategies.fsdp import FSDPStrategy
-    from lightning.pytorch.utilities.rank_zero import rank_zero_info, rank_zero_warn
+    from lightning.pytorch.utilities.rank_zero import rank_zero_warn
 elif module_available("pytorch_lightning"):
     import pytorch_lightning as pl
     from lightning_fabric.plugins import CheckpointIO, ClusterEnvironment
@@ -49,7 +48,7 @@ elif module_available("pytorch_lightning"):
     from lightning_fabric.utilities.types import ReduceOp
     from pytorch_lightning.plugins.precision import Precision
     from pytorch_lightning.strategies.fsdp import FSDPStrategy
-    from pytorch_lightning.utilities.rank_zero import rank_zero_info, rank_zero_warn
+    from pytorch_lightning.utilities.rank_zero import rank_zero_warn
 else:
     raise ModuleNotFoundError("You are missing `lightning` or `pytorch-lightning` package, please install it.")
 
@@ -191,10 +190,6 @@ class HPUFSDPStrategy(FSDPStrategy, HPUParallelStrategy):
         _setup_activation_checkpointing(model, self._activation_checkpointing_kwargs)
 
         return model
-
-    def optimizer_state(self, optimizer: Optimizer) -> Dict[str, torch.Tensor]:
-        rank_zero_info("Optimizer state checkpointing is not enabled yet on HPU.")
-        return super().optimizer_state(optimizer)
 
     def setup(self, trainer: "pl.Trainer") -> None:
         self.model_to_device()