+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for additional dtypes ([#194](https://github.com/Lightning-AI/lightning-Habana/pull/194))
 -
 
 ### Changed